test: gpu unit test
