+++ tests/test_context.py
@@ -1,264 +1,264 @@
-import base64
-import json
-import os
-import unittest
-from unittest import mock
-from loguru import logger
-import yaml
-from kubernetes.client.models.v1_config_map import V1ConfigMap
-from kubernetes.client.models.v1_persistent_volume_claim import V1PersistentVolumeClaim
-from kubernetes.client.models.v1_role import V1Role
-from kubernetes.client.models.v1_role_binding import V1RoleBinding
-from kubernetes.client.models.v1_secret import V1Secret
+# import base64
+# import json
+# import os
+# import unittest
+# from unittest import mock
+# from loguru import logger
+# import yaml
+# from kubernetes.client.models.v1_config_map import V1ConfigMap
+# from kubernetes.client.models.v1_persistent_volume_claim import V1PersistentVolumeClaim
+# from kubernetes.client.models.v1_role import V1Role
+# from kubernetes.client.models.v1_role_binding import V1RoleBinding
+# from kubernetes.client.models.v1_secret import V1Secret
 
-from pycalrissian.context import CalrissianContext
+# from pycalrissian.context import CalrissianContext
 
-os.environ["KUBECONFIG"] = "~/.kube/kubeconfig-t2-dev.yaml"
+# os.environ["KUBECONFIG"] = "~/.kube/kubeconfig-t2-dev.yaml"
 
 
-class TestCalrissianContext(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        logger.info(f"-----\n------------------------------  unit test for test_context.py   ------------------------------\n\n")
-        cls.namespace = "dummy-namespace"
+# class TestCalrissianContext(unittest.TestCase):
+#     @classmethod
+#     def setUpClass(cls):
+#         logger.info(f"-----\n------------------------------  unit test for test_context.py   ------------------------------\n\n")
+#         cls.namespace = "dummy-namespace"
     
-    def test_env(self):
+#     def test_env(self):
 
-        self.assertIsNotNone(os.getenv("KUBECONFIG", None))
+#         self.assertIsNotNone(os.getenv("KUBECONFIG", None))
 
-    def test_core_v1_api(self):
-        logger.info(f"-----\n------------------------------  Testing core_v1_api   ------------------------------\n\n")
-        session = CalrissianContext(
-            namespace=self.namespace, storage_class="dummy", volume_size="1G"
-        )
+#     def test_core_v1_api(self):
+#         logger.info(f"-----\n------------------------------  Testing core_v1_api   ------------------------------\n\n")
+#         session = CalrissianContext(
+#             namespace=self.namespace, storage_class="dummy", volume_size="1G"
+#         )
 
-        self.assertIsNotNone(session.core_v1_api)
+#         self.assertIsNotNone(session.core_v1_api)
         
 
-    def test_rbac_authorization_v1_api(self):
-        logger.info(f"-----\n------------------------------  Testing rbac_authorization_v1_api   ------------------------------\n\n")
-        session = CalrissianContext(
-            namespace=self.namespace, storage_class="dummy", volume_size="1G"
-        )
+#     def test_rbac_authorization_v1_api(self):
+#         logger.info(f"-----\n------------------------------  Testing rbac_authorization_v1_api   ------------------------------\n\n")
+#         session = CalrissianContext(
+#             namespace=self.namespace, storage_class="dummy", volume_size="1G"
+#         )
 
-        self.assertIsNotNone(session.rbac_authorization_v1_api)
+#         self.assertIsNotNone(session.rbac_authorization_v1_api)
 
-    def test_create_namespace(self):
-        logger.info(f"-----\n------------------------------  Testing namespace creation: {self.namespace}   ------------------------------\n\n")
-        session = CalrissianContext(
-            namespace=self.namespace, storage_class="dummy", volume_size="1G"
-        )
+#     def test_create_namespace(self):
+#         logger.info(f"-----\n------------------------------  Testing namespace creation: {self.namespace}   ------------------------------\n\n")
+#         session = CalrissianContext(
+#             namespace=self.namespace, storage_class="dummy", volume_size="1G"
+#         )
 
-        # if session.is_namespace_created():
-        #     session.core_v1_api.delete_namespace(
-        #         name=self.namespace, pretty=True
-        #     )
-        response = session.create_namespace()
+#         # if session.is_namespace_created():
+#         #     session.core_v1_api.delete_namespace(
+#         #         name=self.namespace, pretty=True
+#         #     )
+#         response = session.create_namespace()
 
-        self.assertIsNotNone(response)
+#         self.assertIsNotNone(response)
 
-    def test_create_role_1(self):
-        logger.info(f"-----\n------------------------------  Testing role 1 {self.namespace}   ------------------------------\n\n")
-        session = CalrissianContext(
-            namespace=self.namespace, storage_class="dummy", volume_size="1G"
-        )
+#     def test_create_role_1(self):
+#         logger.info(f"-----\n------------------------------  Testing role 1 {self.namespace}   ------------------------------\n\n")
+#         session = CalrissianContext(
+#             namespace=self.namespace, storage_class="dummy", volume_size="1G"
+#         )
 
-        if not session.is_namespace_created():
-            session.create_namespace()
+#         if not session.is_namespace_created():
+#             session.create_namespace()
 
-        roles = {}
+#         roles = {}
 
-        roles["pod-manager-role"] = {
-            "verbs": ["create", "patch", "delete", "list", "watch"],
-            "role_binding": "pod-manager-default-binding",
-        }
+#         roles["pod-manager-role"] = {
+#             "verbs": ["create", "patch", "delete", "list", "watch"],
+#             "role_binding": "pod-manager-default-binding",
+#         }
 
-        role_name = "pod-manager-role"
+#         role_name = "pod-manager-role"
 
-        response = session.create_role(
-            name=role_name,
-            verbs=roles[role_name]["verbs"],
-            resources=["pods", "pods/log"],
-            api_groups=["*"],
-        )
+#         response = session.create_role(
+#             name=role_name,
+#             verbs=roles[role_name]["verbs"],
+#             resources=["pods", "pods/log"],
+#             api_groups=["*"],
+#         )
 
-        self.assertIsInstance(response, V1Role)
+#         self.assertIsInstance(response, V1Role)
 
-    def test_create_role_binding_1(self):
-        logger.info(f"-----\n------------------------------  Testing role binding {self.namespace}   ------------------------------\n\n")
-        session = CalrissianContext(
-            namespace=self.namespace, storage_class="dummy", volume_size="1G"
-        )
+#     def test_create_role_binding_1(self):
+#         logger.info(f"-----\n------------------------------  Testing role binding {self.namespace}   ------------------------------\n\n")
+#         session = CalrissianContext(
+#             namespace=self.namespace, storage_class="dummy", volume_size="1G"
+#         )
 
-        if not session.is_namespace_created():
-            session.create_namespace()
+#         if not session.is_namespace_created():
+#             session.create_namespace()
 
-        roles = {}
+#         roles = {}
 
-        roles["pod-manager-role"] = {
-            "verbs": ["create", "patch", "delete", "list", "watch"],
-            "role_binding": "pod-manager-default-binding",
-        }
+#         roles["pod-manager-role"] = {
+#             "verbs": ["create", "patch", "delete", "list", "watch"],
+#             "role_binding": "pod-manager-default-binding",
+#         }
 
-        role_name = "pod-manager-role"
+#         role_name = "pod-manager-role"
 
-        response = session.create_role(
-            name=role_name,
-            verbs=roles["pod-manager-role"]["verbs"],
-            resources=["pods", "pods/log"],
-            api_groups=["*"],
-        )
+#         response = session.create_role(
+#             name=role_name,
+#             verbs=roles["pod-manager-role"]["verbs"],
+#             resources=["pods", "pods/log"],
+#             api_groups=["*"],
+#         )
 
-        response = session.create_role_binding(
-            name=roles[role_name]["role_binding"], role=role_name
-        )
+#         response = session.create_role_binding(
+#             name=roles[role_name]["role_binding"], role=role_name
+#         )
 
-        self.assertIsInstance(response, V1RoleBinding)
+#         self.assertIsInstance(response, V1RoleBinding)
 
-    def test_create_volume(self):
-        logger.info(f"-----\n------------------------------  Testing volume {self.namespace}   ------------------------------\n\n")
-        session = CalrissianContext(
-            namespace=self.namespace,
-            storage_class="standard",
-            volume_size="1G",
-        )
+#     def test_create_volume(self):
+#         logger.info(f"-----\n------------------------------  Testing volume {self.namespace}   ------------------------------\n\n")
+#         session = CalrissianContext(
+#             namespace=self.namespace,
+#             storage_class="standard",
+#             volume_size="1G",
+#         )
 
-        if not session.is_namespace_created():
-            session.create_namespace()
+#         if not session.is_namespace_created():
+#             session.create_namespace()
 
-        response = session.create_pvc(
-            name="calrissian-wdir",
-            size=session.volume_size,
-            storage_class=session.storage_class,
-            access_modes=["ReadWriteMany"],
-        )
+#         response = session.create_pvc(
+#             name="calrissian-wdir",
+#             size=session.volume_size,
+#             storage_class=session.storage_class,
+#             access_modes=["ReadWriteMany"],
+#         )
 
-        self.assertIsInstance(response, V1PersistentVolumeClaim)
+#         self.assertIsInstance(response, V1PersistentVolumeClaim)
 
-    def test_configmap_from_dict_as_yaml(self):
-        logger.info(f"-----\n------------------------------  Testing configmap {self.namespace}   ------------------------------\n\n")
-        session = CalrissianContext(
-            namespace=self.namespace,
-            storage_class="standard",
-            volume_size="1G",
-        )
+#     def test_configmap_from_dict_as_yaml(self):
+#         logger.info(f"-----\n------------------------------  Testing configmap {self.namespace}   ------------------------------\n\n")
+#         session = CalrissianContext(
+#             namespace=self.namespace,
+#             storage_class="standard",
+#             volume_size="1G",
+#         )
 
-        if not session.is_namespace_created():
-            session.create_namespace()
+#         if not session.is_namespace_created():
+#             session.create_namespace()
 
-        data = {}
+#         data = {}
 
-        data["cwlVersion"] = "v1.0"
-        data["$graph"] = [
-            {"class": "Workflow", "id": "my_id"},
-            {"class": "CommandLineTool", "id": "my_id"},
-        ]
-        data["key3"] = "value3"
+#         data["cwlVersion"] = "v1.0"
+#         data["$graph"] = [
+#             {"class": "Workflow", "id": "my_id"},
+#             {"class": "CommandLineTool", "id": "my_id"},
+#         ]
+#         data["key3"] = "value3"
 
-        response = session.create_configmap(
-            name="cm-id-yml-cwl",
-            key="from_dict_as_yaml",
-            content=yaml.dump(data),
-        )
+#         response = session.create_configmap(
+#             name="cm-id-yml-cwl",
+#             key="from_dict_as_yaml",
+#             content=yaml.dump(data),
+#         )
 
-        self.assertIsInstance(response, V1ConfigMap)
+#         self.assertIsInstance(response, V1ConfigMap)
 
-    def test_configmap_from_dict_as_json(self):
-        logger.info(f"-----\n------------------------------  Testing configmap {self.namespace}   ------------------------------\n\n")
-        session = CalrissianContext(
-            namespace=self.namespace,
-            storage_class="standard",
-            volume_size="1G",
-        )
+#     def test_configmap_from_dict_as_json(self):
+#         logger.info(f"-----\n------------------------------  Testing configmap {self.namespace}   ------------------------------\n\n")
+#         session = CalrissianContext(
+#             namespace=self.namespace,
+#             storage_class="standard",
+#             volume_size="1G",
+#         )
 
-        if not session.is_namespace_created():
-            session.create_namespace()
+#         if not session.is_namespace_created():
+#             session.create_namespace()
 
-        data = {}
+#         data = {}
 
-        data["key1"] = "value1"
-        data["key2"] = "value2"
-        data["key3"] = "value3"
+#         data["key1"] = "value1"
+#         data["key2"] = "value2"
+#         data["key3"] = "value3"
 
-        response = session.create_configmap(
-            name="cm-id-json",
-            key="from_dict",
-            content=json.dumps(data, indent=4),
-        )
+#         response = session.create_configmap(
+#             name="cm-id-json",
+#             key="from_dict",
+#             content=json.dumps(data, indent=4),
+#         )
 
-        self.assertIsInstance(response, V1ConfigMap)
+#         self.assertIsInstance(response, V1ConfigMap)
 
-    def test_secret_creation(self):
-        logger.info(f"-----\n------------------------------  Testing secret {self.namespace}   ------------------------------\n\n")
-        username = "pippo"
-        password = "pippo"
-        email = "john.doe@me.com"
-        registry = "1ui32139.gra7.container-registry.ovh.net"
+#     def test_secret_creation(self):
+#         logger.info(f"-----\n------------------------------  Testing secret {self.namespace}   ------------------------------\n\n")
+#         username = "pippo"
+#         password = "pippo"
+#         email = "john.doe@me.com"
+#         registry = "1ui32139.gra7.container-registry.ovh.net"
 
-        auth = base64.b64encode(f"{username}:{password}".encode("utf-8")).decode(
-            "utf-8"
-        )
+#         auth = base64.b64encode(f"{username}:{password}".encode("utf-8")).decode(
+#             "utf-8"
+#         )
 
-        secret_config = {
-            "auths": {
-                registry: {
-                    "username": username,
-                    "password": password,
-                    "email": email,
-                    "auth": auth,
-                }
-            }
-        }
+#         secret_config = {
+#             "auths": {
+#                 registry: {
+#                     "username": username,
+#                     "password": password,
+#                     "email": email,
+#                     "auth": auth,
+#                 }
+#             }
+#         }
 
-        session = CalrissianContext(
-            namespace=self.namespace,
-            storage_class="standard",
-            volume_size="1G",
-            image_pull_secrets={"imagePullSecrets": secret_config},
-        )
+#         session = CalrissianContext(
+#             namespace=self.namespace,
+#             storage_class="standard",
+#             volume_size="1G",
+#             image_pull_secrets={"imagePullSecrets": secret_config},
+#         )
 
-        if not session.is_namespace_created():
-            session.create_namespace()
+#         if not session.is_namespace_created():
+#             session.create_namespace()
 
-        response = session.create_image_pull_secret(name="container-rg")
+#         response = session.create_image_pull_secret(name="container-rg")
 
-        self.assertIsInstance(response, V1Secret)
+#         self.assertIsInstance(response, V1Secret)
 
 
-    @mock.patch.object(CalrissianContext, "_get_api_client")
-    @mock.patch.object(CalrissianContext, "_get_core_v1_api")
-    @mock.patch.object(CalrissianContext, "_get_batch_v1_api")
-    @mock.patch.object(CalrissianContext, "_get_rbac_authorization_v1_api")
-    def test_context_from_existing_namespace(
-        self,
-        mock_get_rbac_api,
-        mock_get_batch_api,
-        mock_get_core_api,
-        mock_get_api_client
-    ):
+#     @mock.patch.object(CalrissianContext, "_get_api_client")
+#     @mock.patch.object(CalrissianContext, "_get_core_v1_api")
+#     @mock.patch.object(CalrissianContext, "_get_batch_v1_api")
+#     @mock.patch.object(CalrissianContext, "_get_rbac_authorization_v1_api")
+#     def test_context_from_existing_namespace(
+#         self,
+#         mock_get_rbac_api,
+#         mock_get_batch_api,
+#         mock_get_core_api,
+#         mock_get_api_client
+#     ):
 
-        mock_get_api_client.return_value = mock.Mock()
-        mock_get_batch_api.return_value = mock.Mock()
-        mock_get_rbac_api.return_value = mock.Mock()
+#         mock_get_api_client.return_value = mock.Mock()
+#         mock_get_batch_api.return_value = mock.Mock()
+#         mock_get_rbac_api.return_value = mock.Mock()
 
-        mock_core_v1 = mock.Mock()
-        mock_get_core_api.return_value = mock_core_v1
-        mock_core_v1.read_namespace.return_value = {"metadata": {"name": self.namespace}}
+#         mock_core_v1 = mock.Mock()
+#         mock_get_core_api.return_value = mock_core_v1
+#         mock_core_v1.read_namespace.return_value = {"metadata": {"name": self.namespace}}
         
-        session = CalrissianContext.from_existing_namespace(
-            namespace=self.namespace,
-            storage_class="dummy-storage-class",
-            volume_size="1G"
-        )
+#         session = CalrissianContext.from_existing_namespace(
+#             namespace=self.namespace,
+#             storage_class="dummy-storage-class",
+#             volume_size="1G"
+#         )
 
 
-        self.assertEqual(session.namespace, self.namespace)
-        self.assertEqual(session.storage_class,"dummy-storage-class")
-        self.assertEqual(session.volume_size, "1G")
-        self.assertTrue(session.existing_namespace)
+#         self.assertEqual(session.namespace, self.namespace)
+#         self.assertEqual(session.storage_class,"dummy-storage-class")
+#         self.assertEqual(session.volume_size, "1G")
+#         self.assertTrue(session.existing_namespace)
 
 
-# # if __name__ == "__main__":
-# #     import nose2
+# # # if __name__ == "__main__":
+# # #     import nose2
 
-# #     nose2.main()
+# # #     nose2.main()
