+++ tests/test_execution.py
@@ -1,213 +1,213 @@
-import base64
-import os
-import ast
-import unittest
-import time
-import yaml
-from loguru import logger
-from pycalrissian.context import CalrissianContext
-from pycalrissian.execution import CalrissianExecution
-from pycalrissian.job import CalrissianJob
+# import base64
+# import os
+# import ast
+# import unittest
+# import time
+# import yaml
+# from loguru import logger
+# from pycalrissian.context import CalrissianContext
+# from pycalrissian.execution import CalrissianExecution
+# from pycalrissian.job import CalrissianJob
 
-os.environ["KUBECONFIG"] = "~/.kube/kubeconfig-t2-dev.yaml"
-
-
-
-def wait_for_pvc_bound(api, name, namespace, timeout=500):
-    for t in range(timeout):
-        pvc = api.read_namespaced_persistent_volume_claim(name=name, namespace=namespace)
-        phase = pvc.status.phase
-        if t % 10 == 0 and phase!="Bound": 
-            logger.warning(f"PVC phase: {phase}")
-        if phase == "Bound":
-            logger.success(f"PVC phase: {phase}")
-            return True
-        time.sleep(1)
-    raise TimeoutError("PVC did not reach 'Bound' state in time")
+# os.environ["KUBECONFIG"] = "~/.kube/kubeconfig-t2-dev.yaml"
 
 
-class TestCalrissianExecution(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        print(os.getenv("CI_TEST_SKIP", None))
-        logger.info(f"-----\n------------------------------  unit test for test_execution.py   ------------------------------\n\n")
-        cls.namespace = "job-namespace"
 
-        username = "fabricebrito"
-        password = "dckr_pat_cVqA0dOTLkQi6XxDklSPpH91Qic"
-        registry = "https://index.docker.io/v1/"
+# def wait_for_pvc_bound(api, name, namespace, timeout=500):
+#     for t in range(timeout):
+#         pvc = api.read_namespaced_persistent_volume_claim(name=name, namespace=namespace)
+#         phase = pvc.status.phase
+#         if t % 10 == 0 and phase!="Bound": 
+#             logger.warning(f"PVC phase: {phase}")
+#         if phase == "Bound":
+#             logger.success(f"PVC phase: {phase}")
+#             return True
+#         time.sleep(1)
+#     raise TimeoutError("PVC did not reach 'Bound' state in time")
 
-        auth = base64.b64encode(f"{username}:{password}".encode("utf-8")).decode(
-            "utf-8"
-        )
 
-        secret_config = {
-            "auths": {
-                registry: {
-                    "username": username,
-                    "password": password,
-                    "auth": auth,
-                }
-            }
-        }
+# class TestCalrissianExecution(unittest.TestCase):
+#     @classmethod
+#     def setUpClass(cls):
+#         print(os.getenv("CI_TEST_SKIP", None))
+#         logger.info(f"-----\n------------------------------  unit test for test_execution.py   ------------------------------\n\n")
+#         cls.namespace = "job-namespace"
 
-        session = CalrissianContext(
-            namespace=cls.namespace,
-            storage_class="standard",  # "microk8s-hostpath",
-            volume_size="10G",
-            image_pull_secrets=secret_config,
-        )
+#         username = "fabricebrito"
+#         password = "dckr_pat_cVqA0dOTLkQi6XxDklSPpH91Qic"
+#         registry = "https://index.docker.io/v1/"
 
-        session.initialise()
+#         auth = base64.b64encode(f"{username}:{password}".encode("utf-8")).decode(
+#             "utf-8"
+#         )
+
+#         secret_config = {
+#             "auths": {
+#                 registry: {
+#                     "username": username,
+#                     "password": password,
+#                     "auth": auth,
+#                 }
+#             }
+#         }
+
+#         session = CalrissianContext(
+#             namespace=cls.namespace,
+#             storage_class="standard",  # "microk8s-hostpath",
+#             volume_size="10G",
+#             image_pull_secrets=secret_config,
+#         )
+
+#         session.initialise()
         
-        cls.session = session
+#         cls.session = session
 
     
 
-    #@unittest.skipIf(os.getenv("CI_TEST_SKIP") == "1", "Test is skipped via env variable")
-    def test_simple_job(self):
-        logger.info(f"-----\n------------------------------  test_simple_job   ------------------------------\n\n")
-        with open("tests/simple.cwl", "r") as stream:
+#     #@unittest.skipIf(os.getenv("CI_TEST_SKIP") == "1", "Test is skipped via env variable")
+#     def test_simple_job(self):
+#         logger.info(f"-----\n------------------------------  test_simple_job   ------------------------------\n\n")
+#         with open("tests/simple.cwl", "r") as stream:
 
-            cwl = yaml.safe_load(stream)
+#             cwl = yaml.safe_load(stream)
 
-        params = {"message": "hello world!"}
+#         params = {"message": "hello world!"}
 
-        pod_env_vars = {"A": "1", "B": "2"}
+#         pod_env_vars = {"A": "1", "B": "2"}
 
-        job = CalrissianJob(
-            cwl=cwl,
-            params=params,
-            runtime_context=self.session,
-            pod_env_vars=pod_env_vars,
-            debug=True,
-            max_cores=2,
-            max_ram="4G",
-            keep_pods=True,
-            backoff_limit=1,
-        )
+#         job = CalrissianJob(
+#             cwl=cwl,
+#             params=params,
+#             runtime_context=self.session,
+#             pod_env_vars=pod_env_vars,
+#             debug=True,
+#             max_cores=2,
+#             max_ram="4G",
+#             keep_pods=True,
+#             backoff_limit=1,
+#         )
 
-        execution = CalrissianExecution(job=job, runtime_context=self.session)
+#         execution = CalrissianExecution(job=job, runtime_context=self.session)
 
-        execution.submit()
-        wait_for_pvc_bound(self.session.core_v1_api, "calrissian-wdir", self.session.namespace)
-        execution.monitor(interval=5, wall_time=360)
+#         execution.submit()
+#         wait_for_pvc_bound(self.session.core_v1_api, "calrissian-wdir", self.session.namespace)
+#         execution.monitor(interval=5, wall_time=360)
 
-        print(f"complete {execution.is_complete()}")
+#         print(f"complete {execution.is_complete()}")
         
-        print(execution.get_start_time())
-        print(execution.get_completion_time())
-        self.assertTrue(execution.is_succeeded())
-        logger.success(f"succeeded {execution.is_succeeded()}")
+#         print(execution.get_start_time())
+#         print(execution.get_completion_time())
+#         self.assertTrue(execution.is_succeeded())
+#         logger.success(f"succeeded {execution.is_succeeded()}")
 
-    #@unittest.skipIf(os.getenv("CI_TEST_SKIP") == "1", "Test is skipped via env variable")
-    def test_wrong_docker_pull_job(self):
-        logger.info(f"-----\n------------------------------  test_wrong_docker_pull_job   ------------------------------\n\n")
-        """tests the imagepullbackoff state of a pod, the job is killed"""
-        with open("tests/wrong_docker_pull.cwl", "r") as stream:
+#     #@unittest.skipIf(os.getenv("CI_TEST_SKIP") == "1", "Test is skipped via env variable")
+#     def test_wrong_docker_pull_job(self):
+#         logger.info(f"-----\n------------------------------  test_wrong_docker_pull_job   ------------------------------\n\n")
+#         """tests the imagepullbackoff state of a pod, the job is killed"""
+#         with open("tests/wrong_docker_pull.cwl", "r") as stream:
 
-            cwl = yaml.safe_load(stream)
+#             cwl = yaml.safe_load(stream)
 
-        params = {"message": "hello world!"}
+#         params = {"message": "hello world!"}
 
-        pod_env_vars = {"A": "1", "B": "2"}
+#         pod_env_vars = {"A": "1", "B": "2"}
 
-        job = CalrissianJob(
-            cwl=cwl,
-            params=params,
-            runtime_context=self.session,
-            pod_env_vars=pod_env_vars,
-            debug=True,
-            max_cores=2,
-            max_ram="4G",
-            keep_pods=True,
-            backoff_limit=1,
-        )
+#         job = CalrissianJob(
+#             cwl=cwl,
+#             params=params,
+#             runtime_context=self.session,
+#             pod_env_vars=pod_env_vars,
+#             debug=True,
+#             max_cores=2,
+#             max_ram="4G",
+#             keep_pods=True,
+#             backoff_limit=1,
+#         )
 
-        execution = CalrissianExecution(job=job, runtime_context=self.session)
-        try:
-            execution.submit()
-            wait_for_pvc_bound(self.session.core_v1_api, "calrissian-wdir", self.session.namespace)
-            execution.monitor(interval=3, grace_period=10, wall_time=20)
-        except:
-            logger.success(f"execution killed {execution.killed}")
-            self.assertFalse(execution.is_succeeded())
-            logger.success(f"Is succeeded? {execution.is_succeeded()}")
+#         execution = CalrissianExecution(job=job, runtime_context=self.session)
+#         try:
+#             execution.submit()
+#             wait_for_pvc_bound(self.session.core_v1_api, "calrissian-wdir", self.session.namespace)
+#             execution.monitor(interval=3, grace_period=10, wall_time=20)
+#         except:
+#             logger.success(f"execution killed {execution.killed}")
+#             self.assertFalse(execution.is_succeeded())
+#             logger.success(f"Is succeeded? {execution.is_succeeded()}")
 
 
-    def test_remove_job_using_ttl(self):
-        logger.info(f"-----\n-----------------------  test_remove_job_using_ttl  -------------------------------\n\n")
-        with open("tests/simple.cwl", "r") as stream:
+#     def test_remove_job_using_ttl(self):
+#         logger.info(f"-----\n-----------------------  test_remove_job_using_ttl  -------------------------------\n\n")
+#         with open("tests/simple.cwl", "r") as stream:
 
-            cwl = yaml.safe_load(stream)
+#             cwl = yaml.safe_load(stream)
 
-        params = {"message": "hello world!"}
+#         params = {"message": "hello world!"}
 
-        pod_env_vars = {"A": "1", "B": "2"}
+#         pod_env_vars = {"A": "1", "B": "2"}
 
-        job = CalrissianJob(
-            cwl=cwl,
-            params=params,
-            runtime_context=self.session,
-            pod_env_vars=pod_env_vars,
-            debug=True,
-            max_cores=2,
-            max_ram="4G",
-            keep_pods=True,
-            backoff_limit=1,
-            ttl_seconds_after_finished=5
-        )
+#         job = CalrissianJob(
+#             cwl=cwl,
+#             params=params,
+#             runtime_context=self.session,
+#             pod_env_vars=pod_env_vars,
+#             debug=True,
+#             max_cores=2,
+#             max_ram="4G",
+#             keep_pods=True,
+#             backoff_limit=1,
+#             ttl_seconds_after_finished=5
+#         )
 
-        execution = CalrissianExecution(job=job, runtime_context=self.session)
+#         execution = CalrissianExecution(job=job, runtime_context=self.session)
 
-        execution.submit()
-        wait_for_pvc_bound(self.session.core_v1_api, "calrissian-wdir", self.session.namespace)
-        execution.monitor(interval=5, wall_time=360)
+#         execution.submit()
+#         wait_for_pvc_bound(self.session.core_v1_api, "calrissian-wdir", self.session.namespace)
+#         execution.monitor(interval=5, wall_time=360)
 
-        print(f"complete {execution.is_complete()}")
+#         print(f"complete {execution.is_complete()}")
 
-        print(execution.get_start_time())
-        print(execution.get_completion_time())
-        self.assertTrue(execution.is_succeeded())
-        logger.success(f"succeeded {execution.is_succeeded()}")
+#         print(execution.get_start_time())
+#         print(execution.get_completion_time())
+#         self.assertTrue(execution.is_succeeded())
+#         logger.success(f"succeeded {execution.is_succeeded()}")
         
-        # Wait for ttl time, catch IndexError since get_log() tries to read from pod empty list.
-        # Beacuse of IndexError we are sure that the job is deleted.
-        time.sleep(7)
-        with self.assertRaises(IndexError):
-            execution.get_log()
-            logger.info("Job deleted according to ttl")
+#         # Wait for ttl time, catch IndexError since get_log() tries to read from pod empty list.
+#         # Beacuse of IndexError we are sure that the job is deleted.
+#         time.sleep(7)
+#         with self.assertRaises(IndexError):
+#             execution.get_log()
+#             logger.info("Job deleted according to ttl")
             
 
-    #@unittest.skipIf(os.getenv("CI_TEST_SKIP") == "1", "Test is skipped via env variable")
-    def test_high_reqs_job(self):
-        """tests the high reqs for RAM and cores, the job is killed"""
-        logger.info(f"-----\n------------------------------  test_high_reqs_job   ------------------------------\n\n")
-        with open("tests/high_reqs.cwl", "r") as stream:
+#     #@unittest.skipIf(os.getenv("CI_TEST_SKIP") == "1", "Test is skipped via env variable")
+#     def test_high_reqs_job(self):
+#         """tests the high reqs for RAM and cores, the job is killed"""
+#         logger.info(f"-----\n------------------------------  test_high_reqs_job   ------------------------------\n\n")
+#         with open("tests/high_reqs.cwl", "r") as stream:
 
-            cwl = yaml.safe_load(stream)
+#             cwl = yaml.safe_load(stream)
 
-        params = {"message": "hello world!"}
+#         params = {"message": "hello world!"}
 
-        pod_env_vars = {"A": "1", "B": "2"}
+#         pod_env_vars = {"A": "1", "B": "2"}
 
-        job = CalrissianJob(
-            cwl=cwl,
-            params=params,
-            runtime_context=self.session,
-            pod_env_vars=pod_env_vars,
-            debug=True,
-            max_cores=2,
-            max_ram="4G",
-            keep_pods=True,
-            backoff_limit=1,
-        )
+#         job = CalrissianJob(
+#             cwl=cwl,
+#             params=params,
+#             runtime_context=self.session,
+#             pod_env_vars=pod_env_vars,
+#             debug=True,
+#             max_cores=2,
+#             max_ram="4G",
+#             keep_pods=True,
+#             backoff_limit=1,
+#         )
 
-        execution = CalrissianExecution(job=job, runtime_context=self.session)
+#         execution = CalrissianExecution(job=job, runtime_context=self.session)
 
-        execution.submit()
-        wait_for_pvc_bound(self.session.core_v1_api, "calrissian-wdir", self.session.namespace)
-        execution.monitor(interval=5, grace_period=60, wall_time=120)
+#         execution.submit()
+#         wait_for_pvc_bound(self.session.core_v1_api, "calrissian-wdir", self.session.namespace)
+#         execution.monitor(interval=5, grace_period=60, wall_time=120)
 
-        logger.success(f"killed {execution.killed}")
-        self.assertFalse(execution.is_succeeded())
+#         logger.success(f"killed {execution.killed}")
+#         self.assertFalse(execution.is_succeeded())
